+++ src/lib66/module/regex_replace.c
@@ -81,7 +81,7 @@ void regex_replace(stralloc *filelist, resolve_service_t *res)
                 if ((line[0] != ':') || (get_sep_before(line + 1, ':', '=') < 0))
                     log_die(LOG_EXIT_SYS, "bad format in line: ", line, " of key @infiles field") ;
 
-                r = regex_get_file_name(&filename, line) ;
+                r = (int)regex_get_file_name(&filename, line) ;
                 if (!r)
                     log_dieusys(LOG_EXIT_SYS, "get file name at line: ",  line) ;
